Title: Create the databse file a little earlier.
Description: In the previous version, if a user is running the application
for the first time, the database file was created at a point where the app
asks for a username that is the file is created while the registration is in
process. However, what if the user does not have the database.txt file
and tries to login? This will make the app to open the database.txt file in
read mode, but that file is not avaialbale. The app will crash. This commit
fixes this bug. The file is now created before the registration process starts.